Raise the correct error on classes where when are not defined
Given a class S where the method define is not called, @_enum_hash
will be nil. In order to not raise a undefined method [] for nil
class in const_missing I added the behavior to check for the presence
if the `@_enum_hash` first before trying to access the enums hash.